filesystem restructure complete
