+++ internal/endpoints/computerextensionattributes/computerextensionattributes_crud.go
@@ -1,0 +1,176 @@
+package computerextensionattributes
+
+import (
+	"context"
+	"fmt"
+	"strconv"
+
+	"github.com/deploymenttheory/go-api-sdk-jamfpro/sdk/jamfpro"
+	"github.com/deploymenttheory/terraform-provider-jamfpro/internal/endpoints/common/state"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// resourceJamfProComputerExtensionAttributesCreate is responsible for creating a new Jamf Pro Computer Extension Attribute in the remote system.
+// The function:
+// 1. Constructs the attribute data using the provided Terraform configuration.
+// 2. Calls the API to create the attribute in Jamf Pro.
+// 3. Updates the Terraform state with the ID of the newly created attribute.
+// 4. Initiates a read operation to synchronize the Terraform state with the actual state in Jamf Pro.
+func resourceJamfProComputerExtensionAttributesCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+
+	resource, err := constructJamfProComputerExtensionAttribute(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Computer Extension Attribute: %v", err))
+	}
+
+	var creationResponse *jamfpro.ResourceComputerExtensionAttribute
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutCreate), func() *retry.RetryError {
+		var apiErr error
+		creationResponse, apiErr = client.CreateComputerExtensionAttribute(resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to create Jamf Pro Computer Extension Attribute '%s' after retries: %v", resource.Name, err))
+	}
+
+	d.SetId(strconv.Itoa(creationResponse.ID))
+
+	// checkResourceExists := func(id interface{}) (interface{}, error) {
+	// 	intID, err := strconv.Atoi(id.(string))
+	// 	if err != nil {
+	// 		return nil, fmt.Errorf("error converting ID '%v' to integer: %v", id, err)
+	// 	}
+	// 	return client.GetComputerExtensionAttributeByID(intID)
+	// }
+
+	// _, waitDiags := waitfor.ResourceIsAvailable(ctx, d, "Jamf Pro Computer Extension Attribute", strconv.Itoa(creationResponse.ID), checkResourceExists, time.Duration(common.DefaultPropagationTime)*time.Second)
+	// if waitDiags.HasError() {
+	// 	return waitDiags
+	// }
+
+	return append(diags, resourceJamfProComputerExtensionAttributesRead(ctx, d, meta)...)
+}
+
+// resourceJamfProComputerExtensionAttributesRead is responsible for reading the current state of a Jamf Pro Computer Extension Attribute from the remote system.
+// The function:
+// 1. Fetches the attribute's current state using its ID. If it fails then obtain attribute's current state using its Name.
+// 2. Updates the Terraform state with the fetched data to ensure it accurately reflects the current state in Jamf Pro.
+// 3. Handles any discrepancies, such as the attribute being deleted outside of Terraform, to keep the Terraform state synchronized.
+func resourceJamfProComputerExtensionAttributesRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	client := meta.(*jamfpro.Client)
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	resource, err := client.GetComputerExtensionAttributeByID(resourceIDInt)
+
+	if err != nil {
+		return state.HandleResourceNotFoundError(err, d)
+	}
+
+	return append(diags, updateTerraformState(d, resource)...)
+}
+
+// resourceJamfProComputerExtensionAttributesUpdate is responsible for updating an existing Jamf Pro Computer Extension Attribute on the remote system.
+func resourceJamfProComputerExtensionAttributesUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	// Initialize API client
+	client, ok := meta.(*jamfpro.Client)
+	if !ok {
+		return diag.Errorf("error asserting meta as *client.client")
+	}
+
+	// Initialize variables
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	// Convert resourceID from string to int
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	// Construct the resource object
+	resource, err := constructJamfProComputerExtensionAttribute(d)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to construct Jamf Pro Computer Extension Attribute for update: %v", err))
+	}
+
+	// Update operations with retries
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutUpdate), func() *retry.RetryError {
+		_, apiErr := client.UpdateComputerExtensionAttributeByID(resourceIDInt, resource)
+		if apiErr != nil {
+			return retry.RetryableError(apiErr)
+		}
+		// Successfully updated the resource, exit the retry loop
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to update Jamf Pro Computer Extension Attribute '%s' (ID: %d) after retries: %v", resource.Name, resourceIDInt, err))
+	}
+
+	// Read the resource to ensure the Terraform state is up to date
+	readDiags := resourceJamfProComputerExtensionAttributesRead(ctx, d, meta)
+	if len(readDiags) > 0 {
+		diags = append(diags, readDiags...)
+	}
+
+	return diags
+}
+
+// resourceJamfProComputerExtensionAttributesDelete is responsible for deleting a Jamf Pro Computer Extension Attribute.
+func resourceJamfProComputerExtensionAttributesDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	// Initialize API client
+	client, ok := meta.(*jamfpro.Client)
+	if !ok {
+		return diag.Errorf("error asserting meta as *client.client")
+	}
+
+	// Initialize variables
+	var diags diag.Diagnostics
+	resourceID := d.Id()
+
+	// Convert resourceID from string to int
+	resourceIDInt, err := strconv.Atoi(resourceID)
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("error converting resource ID '%s' to int: %v", resourceID, err))
+	}
+
+	// Use the retry function for the delete operation with appropriate timeout
+	err = retry.RetryContext(ctx, d.Timeout(schema.TimeoutDelete), func() *retry.RetryError {
+		// Attempt to delete by ID
+		apiErr := client.DeleteComputerExtensionAttributeByID(resourceIDInt)
+		if apiErr != nil {
+			// If deleting by ID fails, attempt to delete by Name
+			resourceName := d.Get("name").(string)
+			apiErrByName := client.DeleteComputerExtensionAttributeByNameByID(resourceName)
+			if apiErrByName != nil {
+				// If deletion by name also fails, return a retryable error
+				return retry.RetryableError(apiErrByName)
+			}
+		}
+		// Successfully deleted the resource, exit the retry loop
+		return nil
+	})
+
+	if err != nil {
+		return diag.FromErr(fmt.Errorf("failed to delete Jamf Pro Computer Extension Attribute '%s' (ID: %d) after retries: %v", d.Get("name").(string), resourceIDInt, err))
+	}
+
+	// Clear the ID from the Terraform state as the resource has been deleted
+	d.SetId("")
+
+	return diags
+}
